cleanup ASM transformers
