+++ src/main/java/com/mitchej123/hodgepodge/asm/hooks/early/EarlyASMCallHooks.java
@@ -12,11 +12,11 @@ import cpw.mods.fml.common.discovery.ModCandidate;
 import cpw.mods.fml.common.discovery.asm.ASMModParser;
 import cpw.mods.fml.common.discovery.asm.ModAnnotation;
 
+@SuppressWarnings("unused")
 public class EarlyASMCallHooks {
 
     public static ModContainer build(ASMModParser modParser, File modSource, ModCandidate container) {
         String className = modParser.getASMType().getClassName();
-
         for (ModAnnotation ann : modParser.getAnnotations()) {
             if (modTypes.containsKey(ann.getASMType())) {
                 FMLLog.fine("Identified a mod of type %s (%s) - loading", ann.getASMType(), className);
@@ -28,7 +28,6 @@ public class EarlyASMCallHooks {
                 }
             }
         }
-
         return null;
     }
 }
