+++ src/main/java/com/mitchej123/hodgepodge/asm/transformers/early/EarlyClassTransformer.java
@@ -36,6 +36,7 @@ import org.objectweb.asm.MethodVisitor;
  * order, we will try to minimize the class dependencies on this class, meaning we will not use the usual Configuration
  * class to handle configurations
  */
+@SuppressWarnings("unused")
 public class EarlyClassTransformer implements IClassTransformer {
 
     private static final boolean noNukeBaseMod;
