+++ src/main/java/com/mitchej123/hodgepodge/asm/transformers/thermos/ThermosFurnaceSledgeHammer.java
@@ -26,8 +26,6 @@ public class ThermosFurnaceSledgeHammer implements IClassTransformer {
         if (Common.config.thermosCraftServerClass.equals(transformedName)) {
             LOGGER.info("Patching Thermos or derivative to not break our furnace fix");
             final ClassReader cr = new ClassReader(basicClass);
-            final ClassWriter cw = new ClassWriter(0);
-
             final ClassNode cn = new ClassNode(ASM5);
             cr.accept(cn, 0);
             for (MethodNode m : cn.methods) {
@@ -40,10 +38,10 @@ public class ThermosFurnaceSledgeHammer implements IClassTransformer {
                     m.maxStack = 0;
                 }
             }
+            final ClassWriter cw = new ClassWriter(0);
             cn.accept(cw);
             return cw.toByteArray();
-        } else {
-            return basicClass;
         }
+        return basicClass;
     }
 }
